Backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pdfplumber
alembic
redis
aiokafka
boto3
scikit-learn
transformers
torch
accelerate
python-multipart
pydantic-settings
httpx
pillow
pdf2image
pytesseract
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
gensim
